AIClothModeling/requirements.txt
torch
torchvision
pillow
pandas


